警告なおしたやつ

+++ Team16/Actor/Bulletes/WallReflectionBullet.cpp
@@ -175,14 +175,14 @@ void WallReflectionBullet::hit(BaseObject & other)
 		Sound::getInstance().playSE("burst02");
 	}
 
-	DrawCircle(b_mPosittion.x + 12 / 2, b_mPosittion.y + 16 / 2, b_mCircleSize, GetColor(255, 255, 0), TRUE);
+	DrawCircle((int)b_mPosittion.x + 12 / 2, (int)b_mPosittion.y + 16 / 2, (int)b_mCircleSize, GetColor(255, 255, 0), TRUE);
 }
 
 Vector2 WallReflectionBullet::RotationZ(float ang)
 {
 	Vector2 v;
 	ang = ang + 45.0f;
-	ang = ang * PI / 180.0;
+	ang = ang * PI / 180.0f;
 
 	float sin = sinf(ang);
 	float cos = cosf(ang);
